Split out `PullRequestBuilder` more cleanly (#3384)
