+++ src/Maestro/SubscriptionActorService/NonBatchedPullRequestActorImplementation.cs
@@ -35,8 +35,8 @@ internal class NonBatchedPullRequestActorImplementation : PullRequestActorImplem
         IMergePolicyEvaluator mergePolicyEvaluator,
         ICoherencyUpdateResolver updateResolver,
         BuildAssetRegistryContext context,
-        IRemoteFactory darcFactory,
-        IBasicBarClient barClient,
+        IRemoteFactory remoteFactory,
+        IPullRequestBuilder pullRequestBuilder,
         ILoggerFactory loggerFactory,
         IActionRunner actionRunner,
         IActorProxyFactory<ISubscriptionActor> subscriptionActorFactory,
@@ -47,8 +47,8 @@ internal class NonBatchedPullRequestActorImplementation : PullRequestActorImplem
             mergePolicyEvaluator,
             updateResolver,
             context,
-            darcFactory,
-            barClient,
+            remoteFactory,
+            pullRequestBuilder,
             loggerFactory,
             actionRunner,
             subscriptionActorFactory)
